requirements.txt
Flask==2.2.3
Pillow
geopy
numpy
rasterio
pyimgur